Add git-gutter configuration
1. Enter linum-mode to show the gutter: `M-x linum-mode RET`
2. Make your changes.
3. Save file: `C-x C-s`
4. Move cursor.

You should see `+`, `-` or `=` symbols in your gutter depending on where you
made your changes.

I've found this mode to be a bit unpredictable when trying to get changes to
appear in the gutter. Sometimes the symbols will not disappear even after
reverting the changes and saving the file.

For the moment it works well enough to give valuable feedback without having to
switch to Git's CLI.